Move manifest tests out of the lib.rs file to facilitate generating it.
