+++ flowstdlib/lib.rs
@@ -170,21 +170,3 @@ pub fn get_manifest() -> Result<LibraryManifest> {
 
     Ok(manifest)
 }
-
-#[cfg(test)]
-mod test {
-    use std::path::Path;
-
-    #[test]
-    fn check_manifest() {
-        // check the manifest was created
-        let manifest = Path::new(env!("CARGO_MANIFEST_DIR")).join("manifest.json");
-        assert!(manifest.exists());
-    }
-
-    #[test]
-    fn get_manifest_test() {
-        let manifest = super::get_manifest().expect("Could not get manifest");
-        assert_eq!(manifest.locators.len(), 30);
-    }
-}
