+++ flowstdlib/main.rs
@@ -18,3 +18,21 @@ pub fn main() {
     );
     check_flow_lib_path();
 }
+
+#[cfg(test)]
+mod test {
+    use std::path::Path;
+
+    #[test]
+    fn check_manifest() {
+        // check the manifest was created
+        let manifest = Path::new(env!("CARGO_MANIFEST_DIR")).join("manifest.json");
+        assert!(manifest.exists());
+    }
+
+    #[test]
+    fn get_manifest_test() {
+        let manifest = flowstdlib::get_manifest().expect("Could not get manifest");
+        assert_eq!(manifest.locators.len(), 30);
+    }
+}
